Refactored FractionEquationDetails into 2 seperate classes
