+++ main/src/equation_builders/FractionMultiplyBuilder.java
@@ -2,6 +2,7 @@ package equation_builders;
 
 import equation_entities.Multiply;
 import equation_parameters.FractionEquationDetails;
+import equation_parameters.FractionMultiDivEquationDetails;
 
 public class FractionMultiplyBuilder extends FractionBuilder {
     @Override
@@ -11,6 +12,9 @@ public class FractionMultiplyBuilder extends FractionBuilder {
 
     @Override
     protected void buildOperands(FractionEquationDetails fractionEquationDetails, int seed) {
+        FractionMultiDivEquationDetails fracMultiDivEqnDetails = (FractionMultiDivEquationDetails) fractionEquationDetails;
 
     }
+
+
 }
