+++ main/src/equation_parameters/EquationDetails.java
@@ -3,6 +3,10 @@ package equation_parameters;
 import java.io.Serializable;
 
 public abstract class EquationDetails implements Serializable {
+    private int numOfEquations;
+    private char operator;
+    private boolean negAllowed;
+
     public int getNumOfEquations() {
         return numOfEquations;
     }
@@ -26,9 +30,5 @@ public abstract class EquationDetails implements Serializable {
     public void setNegAllowed(boolean negAllowed) {
         this.negAllowed = negAllowed;
     }
-
-    private int numOfEquations;
-    private char operator;
-    private boolean negAllowed;
 
 }
