+++ main/src/equation_parameters/FractionMultiDivEquationDetails.java
@@ -1,0 +1,23 @@
+package equation_parameters;
+
+public class FractionMultiDivEquationDetails extends FractionEquationDetails{
+    private int[] ansDenominatorRange;
+    private int complexity;
+
+    public int[] getAnsDenominatorRange() {
+        return ansDenominatorRange;
+    }
+
+    public void setAnsDenominatorRange(int[] ansDenominatorRange) {
+        this.ansDenominatorRange = ansDenominatorRange;
+    }
+
+    public int getComplexity() {
+        return complexity;
+    }
+
+    public void setComplexity(int complexity) {
+        this.complexity = complexity;
+    }
+
+}
